refactor(games): 좋아요 캐시 동기화와 공통 헬퍼 분리
